refactor(ST-2056): add 'vc' prefix to base components and antoher that names as reserved el names
