+++ client-app/components/atoms/alert/vc-alert.vue
