+++ client-app/components/atoms/card-skeleton/vc-card-skeleton.vue
