+++ client-app/components/atoms/chip/vc-chip.vue
