+++ client-app/components/atoms/empty-view/vc-empty-view.vue
