+++ client-app/components/atoms/image/vc-image.vue
