+++ client-app/components/atoms/index.ts
@@ -1,23 +1,23 @@
-export { default as VcAlert } from "./alert/alert.vue";
-export { default as VcButton } from "./button/button.vue";
-export { default as VcBreadcrumbs } from "./breadcrumbs/breadcrumbs.vue";
-export { default as VcCard } from "./card/card.vue";
-export { default as VcCheckbox } from "./checkbox/checkbox.vue";
-export { default as VcChip } from "./chip/chip.vue";
-export { default as VcInput } from "./input/input.vue";
-export { default as VcImage } from "./image/image.vue";
-export { default as VcInfinityScrollLoader } from "./infinity-scroll-loader/infinity-scroll-loader.vue";
-export { default as VcMarkdownRender } from "./markdown-render/markdown-render.vue";
-export { default as VcPagination } from "./pagination/pagination.vue";
-export { default as VcPriceDisplay } from "./price-display/price-display.vue";
-export { default as VcCardSkeleton } from "./card-skeleton/card-skeleton.vue";
-export { default as VcSection } from "./section/section.vue";
-export { default as VcSelect } from "./select/select.vue";
-export { default as VcRadioButton } from "./radio-button/radio-button.vue";
-export { default as VcActionInput } from "./action-input/action-input.vue";
-export { default as VcTabs } from "./tabs/tabs.vue";
-export { default as VcScrollTopButton } from "./scroll-top-button/scroll-top-button.vue";
-export { default as VcDateSelector } from "./date-selector/date-selector.vue";
-export { default as VcPopupSidebar } from "./popup-sidebar/popup-sidebar.vue";
-export { default as VcTextArea } from "./text-area/text-area.vue";
-export { default as VcEmptyView } from "./empty-view/empty-view.vue";
+export { default as VcAlert } from "./alert/vc-alert.vue";
+export { default as VcButton } from "./button/vc-button.vue";
+export { default as VcBreadcrumbs } from "./breadcrumbs/vc-breadcrumbs.vue";
+export { default as VcCard } from "./card/vc-card.vue";
+export { default as VcCheckbox } from "./checkbox/vc-checkbox.vue";
+export { default as VcChip } from "./chip/vc-chip.vue";
+export { default as VcInput } from "./input/vc-input.vue";
+export { default as VcImage } from "./image/vc-image.vue";
+export { default as VcInfinityScrollLoader } from "./infinity-scroll-loader/vc-infinity-scroll-loader.vue";
+export { default as VcMarkdownRender } from "./markdown-render/vc-markdown-render.vue";
+export { default as VcPagination } from "./pagination/vc-pagination.vue";
+export { default as VcPriceDisplay } from "./price-display/vc-price-display.vue";
+export { default as VcCardSkeleton } from "./card-skeleton/vc-card-skeleton.vue";
+export { default as VcSection } from "./section/vc-section.vue";
+export { default as VcSelect } from "./select/vc-select.vue";
+export { default as VcRadioButton } from "./radio-button/vc-radio-button.vue";
+export { default as VcActionInput } from "./action-input/vc-action-input.vue";
+export { default as VcTabs } from "./tabs/vc-tabs.vue";
+export { default as VcScrollTopButton } from "./scroll-top-button/vc-scroll-top-button.vue";
+export { default as VcDateSelector } from "./date-selector/vc-date-selector.vue";
+export { default as VcPopupSidebar } from "./popup-sidebar/vc-popup-sidebar.vue";
+export { default as VcTextArea } from "./text-area/vc-text-area.vue";
+export { default as VcEmptyView } from "./empty-view/vc-empty-view.vue";
