+++ client-app/components/atoms/infinity-scroll-loader/vc-infinity-scroll-loader.vue
