+++ client-app/components/atoms/popup-sidebar/vc-popup-sidebar.vue
