+++ client-app/components/atoms/price-display/vc-price-display.vue
