+++ client-app/components/atoms/tabs/vc-tabs.vue
