+++ client-app/components/molecules/carousel/vc-carousel.vue
