+++ client-app/components/molecules/empty-page/vc-empty-page.vue
