+++ client-app/components/molecules/image-gallery/vc-image-gallery.vue
