+++ client-app/components/molecules/index.ts
@@ -1,4 +1,4 @@
-export { default as VcCarousel } from "./carousel/carousel.vue";
-export { default as VcImageGallery } from "./image-gallery/image-gallery.vue";
-export { default as VcItemPrice } from "./item-price/item-price.vue";
-export { default as VcEmptyPage } from "./empty-page/empty-page.vue";
+export { default as VcCarousel } from "./carousel/vc-carousel.vue";
+export { default as VcImageGallery } from "./image-gallery/vc-image-gallery.vue";
+export { default as VcItemPrice } from "./item-price/vc-item-price.vue";
+export { default as VcEmptyPage } from "./empty-page/vc-empty-page.vue";
