+++ client-app/components/molecules/item-price/vc-item-price.vue
