+++ client-app/components/organisms/index.ts
@@ -1,5 +1,5 @@
-export { default as VcPopup } from "./popup/popup.vue";
-export { default as VcTable } from "./table/table.vue";
+export { default as VcPopup } from "./popup/vc-popup.vue";
+export { default as VcTable } from "./table/vc-table.vue";
 
 // FIXME: Remove private exports
 export { default as TableStatusBadge } from "./table/_internal/table-status-badge/table-status-badge.vue";
