+++ client-app/components/organisms/table/vc-table.vue
