+++ client-app/shared/layout/components/currency-selector/currency-selector.vue
