+++ client-app/shared/layout/components/footer/vc-footer.vue
