+++ client-app/shared/layout/components/header/vc-header.vue
