+++ client-app/shared/layout/components/index.ts
@@ -1,8 +1,8 @@
 export { default as BackButtonInHeader } from "./back-button-in-header/back-button-in-header.vue";
-export { default as Footer } from "./footer/footer.vue";
-export { default as Header } from "./header/header.vue";
+export { default as VcFooter } from "./footer/vc-footer.vue";
+export { default as VcHeader } from "./header/vc-header.vue";
 export { default as SearchBar } from "./search-bar/search-bar.vue";
-export { default as CurrencySelector } from "./сurrency-selector/сurrency-selector.vue";
+export { default as CurrencySelector } from "./currency-selector/currency-selector.vue";
 export { default as TwoColumn } from "./two-column.vue";
 export { default as LanguageSelector } from "./language-selector/language-selector.vue";
 export { default as NestedMobileHeader } from "./nested-mobile-header/nested-mobile-header";
