+++ package.json
@@ -41,7 +41,7 @@
     "moment": "^2.29.1",
     "swiper": "^8.0.6",
     "vee-validate": "^4.5.6",
-    "vue": "3.2.31",
+    "vue": "^3.2.31",
     "vue-i18n": "^9.1.10",
     "vue-router": "^4.0.12",
     "yup": "^0.32.11"
